requirements.txt
streamlit>=1.37.0
supabase>=2.3.0
python-dotenv>=1.0.0
plotly>=5.18.0
streamlit-sortables>=0.2.0
